Added JS and my user input form to 'signin.html' page, removed old form.
Tried to create a button that would call calcBAC but it is coming up as
NaN (not a number), so will ask James tomorrow.

Sorry, thought I was working on my own branch but for some reason
gitHub won’t let me push changes to that, only to master…

Anyway, if you could have a look at styling this when you’re ready and
letting me know if you need any clarification etc, these should be the
correct input fields now. Don’t worry about the p tags that are
confirming the input with the user, they are just there for js testing
at the moment.

Sorry about ruining the nice layout, haven’t written much CSS for this,
just functionality