Added functional drink adding test to index.html page
Test is just functional, minimal styling applied. It uses the same
methods we will hopefully apply for the final product.

Please note that as the index.html and sign-in.html pages are currently
separate, the user inputs form sign-in do not yet carry over to the
tracker page, meaning that it does not yet use the user weight and
gender, but the presets of 60kg and male.

The test currently adds 1 hour per drink entered, and the drink types
do not reflect real SD numbers but rather just values of 1, 2 and 3 SDs
for testing purposes. Feel free to try it out and to check out the
code.

I have also added the ability to change the previous decimal returns
for SoberIn time to hours and mins :)