Form validation, Profile pic input test, Circle visualiser testing page
Form validation built into sign-in page, testing input type=“file” to
see if we want to look at users entering a photo for a profile pic type
of thing.

Also a separate page for testing the circle animation has been started,
but still needs work as it is just seeing if the animation method works
currently.