Added InstantCalc + Revoked timer introduction + Added secondary arc colour for people over 0.1 BAC
Added a new instantCalc.html page for people who need to quickly
estimate BAC without having tracked it the whole night.

Also fixed the way that the grey circle image was displaying in front
of the drawn circle by making it a layer on the canvas. This makes the
animation and visual look much nicer.

Also took back changes that were breaking the code due to a failed
attempt to introduce a timer.

Introduced a new colour to the arc for people who go over 0.1 BAC to
provide the user with better visual information. This applies on both
instantCalc and tracker pages. Currently working on a better way to do
it so that it only starts the new colour after exactly one circle, as
it currently starts the new colour from the start of the drink that
pushes the user’s BAC over 0.1.