FINALLY fixed the animation for 2nd rev, and added multiple canvas layers
Woohoo! Fixed the animation for crossing over between the first and
second revolution and have added multiple canvas layers so the second
circle in red is on separate layer to the other colours underneath it,
effectively preserving the data :)

Moving forwards and looking at getting it to count down with time ASAP!

Also cleaned up the JS and added more comments for easy reading, plus
moved the drinks Array to a separate file to declutter the main JS and
make it easier to edit.